logs big refactoring work in progress 4
