+++ android/src/main/java/io/parity/signer/screens/settings/SettingsScreen.kt
@@ -116,7 +116,11 @@ fun SettingsScreenSubgraph(
 				navController.popBackStack(SettingsScreenSubgraph.home, false)
 			}
 		}
-		logsNavigationSubgraph(SettingsScreenSubgraph.logs)
+		logsNavigationSubgraph(
+			SettingsScreenSubgraph.logs,
+			rootNavigator,
+			navController
+		)
 	}
 }
